LayoutTests/editing/editability/design-mode-does-not-inherit-across-frames-expected.txt

This tests that design mode does not inherit across frame boundaries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.contentDocument.designMode is "off"
PASS document.designMode = "on"; iframe.contentDocument.designMode is "off"
PASS document.designMode = "off"; iframe.contentDocument.designMode is "off"
PASS iframe.contentDocument.designMode = "on"; iframe.contentDocument.designMode is "on"
PASS successfullyParsed is true

TEST COMPLETE
